feat(ticketing): extract filter form and allow partial date filters (#112615)
- Add FILTER_FIELDS constant
- Allow date filters to accept just "from" or "to" values (previously
required both)
- Make filter composable: extract form logic into ticket-filter-form
component
