+++ app/[tenant]/[workspace]/(subapps)/ticketing/common/constants/index.ts
@@ -66,6 +66,15 @@ export const FIELDS = {
   TASK_END_DATE: 'taskEndDate',
 };
 
+export const FILTER_FIELDS = [
+  FIELDS.PRIORITY,
+  FIELDS.STATUS,
+  FIELDS.UPDATED_ON,
+  FIELDS.CREATED_BY,
+  FIELDS.MANAGED_BY,
+  FIELDS.ASSIGNMENT,
+];
+
 export const UPDATABLE_FIELDS = [
   FIELDS.PRIORITY,
   FIELDS.CATEGORY,
