+++ app/[tenant]/[workspace]/(subapps)/ticketing/common/ui/components/ticket-filter-form/index.ts
@@ -1,0 +1,1 @@
+export {TicketFilterForm, type FilterFormProps} from './ticket-filter-form';
