+++ app/[tenant]/[workspace]/(subapps)/ticketing/common/utils/validators.ts
@@ -74,63 +74,52 @@ export const ChildTicketSchema = z.object({
   ),
 });
 
-export const FilterSchema = z.object({
+const dateFilterSchema = z
+  .tuple([z.string().optional(), z.string().optional()])
+  .superRefine((data, ctx) => {
+    const [start, end] = data;
+    if (!start || !end) return;
+    const [startDate, endDate] = [start, end].map(d => new Date(d).getTime());
+    if (startDate >= endDate) {
+      ctx.addIssue({
+        code: 'custom',
+        message: 'Start date must be earlier than End date.',
+      });
+    }
+  });
+
+export const TicketFilterSchema = z.object({
   createdBy: z.array(z.string()).optional(),
   priority: z.array(z.string()).optional(),
   status: z.array(z.string()).optional(),
   category: z.array(z.string()).optional(),
-  updatedOn: z
-    .tuple([z.string().optional(), z.string().optional()])
-    .superRefine((data, ctx) => {
-      const [start, end] = data;
-      if (!start && !end) return;
-      if (!start) {
-        ctx.addIssue({
-          code: 'custom',
-          message: 'Start date is required.',
-          path: [0],
-        });
-        return;
-      }
-      if (!end) {
-        ctx.addIssue({
-          code: 'custom',
-          message: 'End date is required.',
-          path: [1],
-        });
-        return;
-      }
-      const [startDate, endDate] = [start, end].map(d => new Date(d).getTime());
-      if (startDate >= endDate) {
-        ctx.addIssue({
-          code: 'custom',
-          message: 'Start date must be earlier than End date.',
-        });
-      }
-    })
-    .optional(),
+  updatedOn: dateFilterSchema.optional(),
   myTickets: z.boolean().optional(),
   managedBy: z.array(z.string()).optional(),
   assignment: z.number().nullable().optional(),
 });
 
-export const EncodedFilterSchema = FilterSchema.partial().transform(arg => {
-  const filter = Object.fromEntries(
-    Object.entries(arg).filter(([_, value]) => {
-      if (Array.isArray(value)) {
-        return value.length && value.every(v => v != undefined && v != ''); // remove empty arrays and arrays with empty values
-      }
-      if (typeof value === 'boolean') {
-        return value; // remove false
-      }
-      if (value == null) return false; // remove null and undefined
-      return true;
-    }),
-  ) as Omit<Partial<z.infer<typeof FilterSchema>>, 'updatedOn'> & {
-    updatedOn?: [string, string];
-  };
-  if (!Object.keys(filter).length) return null; // remove empty object
-  return filter;
-});
+export const EncodedTicketFilterSchema = TicketFilterSchema.partial().transform(
+  arg => {
+    const filter = Object.fromEntries(
+      Object.entries(arg).filter(([_, value]) => {
+        if (Array.isArray(value)) {
+          return value.length && value.some(v => v != undefined && v != ''); // remove empty arrays
+        }
+        if (typeof value === 'boolean') {
+          return value; // remove false
+        }
+        if (value == null) return false; // remove null and undefined
+        return true;
+      }),
+    ) as Omit<Partial<z.infer<typeof TicketFilterSchema>>, 'updatedOn'> & {
+      updatedOn?: [string, string];
+    };
+    if (!Object.keys(filter).length) return null; // remove empty object
+    return filter;
+  },
+);
 
-export type EncodedFilter = Expand<z.infer<typeof EncodedFilterSchema>>;
+export type EncodedTicketFilter = Expand<
+  z.infer<typeof EncodedTicketFilterSchema>
+>;
