+++ utils/orm.ts
@@ -1,4 +1,4 @@
-import {Entity, WhereOptions} from '@goovee/orm';
+import type {Entity, WhereOptions, DateFilter} from '@goovee/orm';
 import {isEmpty} from 'lodash-es';
 
 export function and<T extends Entity>(
@@ -19,4 +19,15 @@ export function or<T extends Entity>(
     .filter(f => !isEmpty(f)) as WhereOptions<T>[];
   if (filtered.length === 0) return;
   return {OR: filtered};
+}
+
+export function getDateFilter(
+  dates: [string | undefined, string | undefined],
+): DateFilter | null {
+  const startDate = dates[0];
+  const endDate = dates[1];
+  if (startDate && endDate) return {between: [startDate, endDate]};
+  if (startDate) return {ge: startDate};
+  if (endDate) return {le: endDate};
+  return null;
 }
